requirements.txt
pandas==1.4.3
SQLAlchemy==1.4.39
psycopg2-binary==2.9.3
matplotlib==3.5.2
PyYAML==6.0
requests==2.28.1
Jinja2==3.1.2
